refactor: move user::user types from lib to wire crate
