+++ lib/src/user/user.rs
@@ -1,77 +1,13 @@
 use crate::common::{request, request_bare, SerializableNone};
 use crate::error::ApiError;
-use crate::user::ProjectMinimal;
 use anyhow::Context;
-use lrzcc_wire::common::{is_false, is_true};
+use lrzcc_wire::user::{
+    User, UserCreateData, UserCreated, UserDetailed, UserImport, UserModifyData,
+};
 use reqwest::blocking::Client;
 use reqwest::Url;
 use reqwest::{Method, StatusCode};
-use serde::{Deserialize, Serialize};
-use std::fmt::Display;
 use std::rc::Rc;
-use tabled::Tabled;
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct User {
-    pub id: u32,
-    pub name: String,
-    pub openstack_id: String, // UUIDv4 without dashes
-    pub project: u32,
-    pub project_name: String,
-    pub role: u32,
-    pub is_staff: bool,
-    pub is_active: bool,
-}
-
-impl Display for User {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!("User(id={}, name={}", self.id, self.name))
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct UserMinimal {
-    pub id: u32,
-    pub name: String,
-}
-
-impl Display for UserMinimal {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!("User(id={}, name={}", self.id, self.name))
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct UserDetailed {
-    pub id: u32,
-    pub name: String,
-    pub openstack_id: String, // UUIDv4 without dashes
-    pub project: ProjectMinimal,
-    pub project_name: String,
-    pub role: u32,
-    pub is_staff: bool,
-    pub is_active: bool,
-}
-
-// TODO can we merge this with UserDetailed via some enum
-// in the project field
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct UserCreated {
-    pub id: u32,
-    pub name: String,
-    pub openstack_id: String, // UUIDv4 without dashes
-    pub project: u32,
-    pub project_name: String,
-    pub role: u32,
-    pub is_staff: bool,
-    pub is_active: bool,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct UserImport {
-    pub new_project_count: u32,
-    pub new_user_count: u32,
-}
 
 pub struct UserApi {
     pub url: String,
@@ -130,34 +66,6 @@ impl UserListRequest {
     }
 }
 
-#[derive(Clone, Debug, Serialize)]
-struct UserCreateData {
-    name: String,
-    openstack_id: String, // UUIDv4
-    // TODO can't this be optional?
-    project: u32,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    // this could be an enum right
-    role: Option<u32>,
-    #[serde(skip_serializing_if = "is_false")]
-    is_staff: bool,
-    #[serde(skip_serializing_if = "is_true")]
-    is_active: bool,
-}
-
-impl UserCreateData {
-    fn new(name: String, openstack_id: String, project: u32) -> Self {
-        Self {
-            name,
-            openstack_id,
-            project,
-            role: None,
-            is_staff: false,
-            is_active: true,
-        }
-    }
-}
-
 pub struct UserCreateRequest {
     url: String,
     client: Rc<Client>,
@@ -203,38 +111,6 @@ impl UserCreateRequest {
             Some(&self.data),
             StatusCode::CREATED,
         )
-    }
-}
-
-#[derive(Clone, Debug, Serialize)]
-struct UserModifyData {
-    id: u32,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    openstack_id: Option<String>, // UUIDv4
-    #[serde(skip_serializing_if = "Option::is_none")]
-    project: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    role: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    is_staff: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    is_active: Option<bool>,
-}
-
-impl UserModifyData {
-    fn new(id: u32) -> Self {
-        Self {
-            id,
-            name: None,
-            openstack_id: None,
-            project: None,
-            role: None,
-            is_staff: None,
-            is_active: None,
-        }
     }
 }
